% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat_nb.R
\name{chat_nb}
\alias{chat_nb}
\title{Calculate Upper Frechet-Hoeffding Correlation Bound: Negative Binomial - Normal Variables}
\usage{
chat_nb(size, prob, mu = NULL, n_unif = 10000, seed = 1234)
}
\arguments{
\item{size}{a vector of size parameters for the Negative Binomial variables (see \code{\link[stats]{dnbinom}})}

\item{prob}{a vector of success probability parameters}

\item{mu}{a vector of mean parameters (*Note: either \code{prob} or \code{mu} should be supplied for all Negative Binomial variables,
not a mixture; default = NULL)}

\item{n_unif}{the number of uniform random numbers to generate in calculating the bound (default = 10000)}

\item{seed}{the seed used in random number generation (default = 1234)}
}
\value{
A scalar equal to the correlation upper bound.
}
\description{
This function calculates the upper Frechet-Hoeffding bound on the correlation between a Negative Binomial variable
    and the normal variable used to generate it.  It is used in \code{\link[SimMultiCorrData]{findintercorr_cat_nb}}
    and \code{\link[SimMultiCorrData]{findintercorr_cont_nb}} in calculating the intermediate MVN correlations.  This extends
    the method of Amatya & Demirtas (2015) to Negative Binomial variables.  This function would not ordinarily be called directly by the user.
}
\references{
Amatya A & Demirtas H (2015). Simultaneous generation of multivariate mixed data with Poisson and normal marginals.
    Journal of Statistical Computation and Simulation 85(15): 3129-39.

Yahav I & Shmueli G (2012). On Generating Multivariate Poisson Data in Management Science Applications. Applied Stochastic
    Models in Business and Industry, 28(1): 91-102. \doi{10.1002/asmb.901}.

Demirtas H & Hedeker D (2011). A practical way for computing approximate lower and upper correlation bounds.
    American Statistician 65(2): 104-109.

Hoeffding W. Scale-invariant correlation theory. In: Fisher NI, Sen PK, editors. The collected works of Wassily Hoeffding.
    New York: Springer-Verlag; 1994. p. 57-107.

Frechet M.  Sur les tableaux de correlation dont les marges sont donnees.  Ann. l'Univ. Lyon SectA.  1951;14:53-77.
}
\seealso{
\code{\link[SimMultiCorrData]{findintercorr_cat_nb}}, \code{\link[SimMultiCorrData]{findintercorr_cont_nb}},
    \code{\link[SimMultiCorrData]{findintercorr}}
}
\keyword{1}
\keyword{Binomial,}
\keyword{Negative}
\keyword{correlation,}
\keyword{method}
