% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findintercorr_cont_pois.R
\name{findintercorr_cont_pois}
\alias{findintercorr_cont_pois}
\title{Calculate Intermediate MVN Correlation for Continuous - Poisson Variables: Method 1}
\usage{
findintercorr_cont_pois(method, constants, rho_cont_pois, lam, nrand = 1e+05,
  seed = 1234)
}
\arguments{
\item{method}{the method used to generate the \code{k_cont} continuous variables.  "Fleishman" uses a third-order polynomial transformation
and "Polynomial" uses Headrick's fifth-order transformation.}

\item{constants}{a matrix with \code{k_cont} rows, each a vector of constants c0, c1, c2, c3 (if \code{method} = "Fleishman") or
c0, c1, c2, c3, c4, c5 (if \code{method} = "Polynomial"), like that returned by \code{\link[SimMultiCorrData]{find_constants}}}

\item{rho_cont_pois}{a \code{k_cont x k_pois} matrix of target correlations among continuous and Poisson variables}

\item{lam}{a vector of lambda (> 0) constants for the Poisson variables (see \code{\link[stats]{dpois}})}

\item{nrand}{the number of random numbers to generate in calculating the bound (default = 10000)}

\item{seed}{the seed used in random number generation (default = 1234)}
}
\value{
a \code{k_cont x k_pois} matrix whose rows represent the \code{k_cont} continuous variables and columns represent the
    \code{k_pois} Poisson variables
}
\description{
This function calculates a \code{k_cont x k_pois} intermediate matrix of correlations for the \code{k_cont} continuous and
    \code{k_pois} Poisson variables. It extends the method of Amatya & Demirtas (2015) to continuous variables generated using
    Headrick's fifth-order polynomial transformation.  Here, the intermediate correlation between Z1 and Z2 (where Z1 is
    the standard normal variable transformed using Headrick's fifth-order or Fleishman's third-order method to produce a
    continuous variable Y1, and Z2 is the standard normal variable used to generate a Poisson variable via the inverse cdf method) is
    calculated by dividing the target correlation by a correction factor.  The correction factor is the product of the
    upper Frechet-Hoeffding bound on the correlation between a Poisson variable and the normal variable used to generate it
    (see \code{\link[SimMultiCorrData]{chat_pois}}) and the power method correlation (described in Headrick & Kowalchuk, 2007)
    between Y1 and Z1.  The function is used in \code{\link[SimMultiCorrData]{findintercorr}} and
    \code{\link[SimMultiCorrData]{rcorrvar}}.  This function would not ordinarily be called by the user.
}
\references{
Amatya A & Demirtas H (2015). Simultaneous generation of multivariate mixed data with Poisson and normal marginals.
    Journal of Statistical Computation and Simulation, 85(15): 3129-39.

Yahav I & Shmueli G (2012). On Generating Multivariate Poisson Data in Management Science Applications. Applied Stochastic
    Models in Business and Industry, 28(1): 91-102. \doi{10.1002/asmb.901}.

Demirtas H & Hedeker D (2011). A practical way for computing approximate lower and upper correlation bounds.
    American Statistician, 65(2): 104-109.

Hoeffding W. Scale-invariant correlation theory. In: Fisher NI, Sen PK, editors. The collected works of Wassily Hoeffding.
    New York: Springer-Verlag; 1994. p. 57-107.

Frechet M.  Sur les tableaux de correlation dont les marges sont donnees.  Ann. l'Univ. Lyon SectA.  1951;14:53-77.

Headrick TC (2002). Fast Fifth-order Polynomial Transforms for Generating Univariate and Multivariate
    Non-normal Distributions. Computational Statistics & Data Analysis 40(4):685-711
    (\href{http://www.sciencedirect.com/science/article/pii/S0167947302000725}{ScienceDirect})

Fleishman AI (1978). A Method for Simulating Non-normal Distributions. Psychometrika, 43, 521-532.

Headrick TC, Sawilowsky SS (1999). Simulating Correlated Non-normal Distributions: Extending the Fleishman Power
    Method. Psychometrika, 64, 25-35.

Headrick TC, Kowalchuk RK (2007). The Power Method Transformation: Its Probability Density Function, Distribution
    Function, and Its Further Use for Fitting Data. Journal of Statistical Computation and Simulation, 77, 229-249.

Headrick TC, Sheng Y, & Hodis FA (2007). Numerical Computing and Graphics for the Power Method Transformation Using
    Mathematica. Journal of Statistical Software, 19(3), 1 - 17. \doi{10.18637/jss.v019.i03}.

Headrick TC (2004). On Polynomial Transformations for Simulating Multivariate Nonnormal Distributions.
    Journal of Modern Applied Statistical Methods, 3, 65-71.
}
\seealso{
\code{\link[SimMultiCorrData]{chat_pois}}, \code{\link[SimMultiCorrData]{power_norm_corr}},
    \code{\link[SimMultiCorrData]{find_constants}},
    \code{\link[SimMultiCorrData]{findintercorr}}, \code{\link[SimMultiCorrData]{rcorrvar}}
}
\keyword{1}
\keyword{Fleishman,}
\keyword{Headrick,}
\keyword{Poisson,}
\keyword{continuous,}
\keyword{correlation,}
\keyword{intermediate,}
\keyword{method}
