% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fleish.R
\name{fleish}
\alias{fleish}
\title{Fleishman's Third-Order Polynomial Transformation Equations}
\usage{
fleish(c, a)
}
\arguments{
\item{c}{a vector of constants c1, c2, c3; note that \code{\link[SimMultiCorrData]{find_constants}} returns c0, c1, c2, c3}

\item{a}{a vector c(skewness, standardized kurtosis)}
}
\value{
a list of length 3; if the constants satisfy the equations, returns 0 for all list elements
}
\description{
This function contains Fleishman's third-order polynomial transformation equations.  It is used in
    \code{\link[SimMultiCorrData]{find_constants}} to find the constants c1, c2, and c3 (c0 = -c2) that satisfy the
    equations given skewness and standardized kurtosis values.  It can be used to verify a set of constants satisfy
    the equations.  Note that there exist solutions that yield invalid power method pdfs (see
    \code{\link[SimMultiCorrData]{power_norm_corr}}, \code{\link[SimMultiCorrData]{pdf_check}}).
    This function would not ordinarily be called by the user.
}
\examples{
# Laplace Distribution
fleish(c = c(0.782356, 0, 0.067905), a = c(0, 3))
}
\references{
Fleishman AI (1978). A Method for Simulating Non-normal Distributions. Psychometrika, 43, 521-532.

Headrick TC, Sawilowsky SS (1999). Simulating Correlated Non-normal Distributions: Extending the Fleishman Power
Method. Psychometrika, 64, 25-35.

Headrick TC (2004). On Polynomial Transformations for Simulating Multivariate Nonnormal Distributions.
Journal of Modern Applied Statistical Methods, 3, 65-71.
}
\seealso{
\code{\link[SimMultiCorrData]{poly}}, \code{\link[SimMultiCorrData]{power_norm_corr}},
    \code{\link[SimMultiCorrData]{pdf_check}}, \code{\link[SimMultiCorrData]{find_constants}}
}
\keyword{Fleishman}
\keyword{constants,}
