% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_cat.R
\name{var_cat}
\alias{var_cat}
\title{Calculate Variance of Binary or Ordinal Variable}
\usage{
var_cat(marginal, support)
}
\arguments{
\item{marginal}{a vector of cumulative probabilities defining the marginal distribution of the variable;
if the variable can take r values, the vector will contain r - 1 probabilities (the r-th is assumed to be 1)}

\item{support}{a vector of containing the ordered support values}
}
\value{
A scalar equal to the variance
}
\description{
This function calculates the variance of a binary or ordinal (r > 2 categories) variable.  It uses the
    formula given by Olsson, Drasgow, & Dorans (1982) in describing polyserial and point-polyserial correlations.  The
    function is used to find intercorrelations involving ordinal variables or variables that are treated as ordinal
    (i.e. count variables in the Barbiero & Ferrari, 2015 method used in \code{\link[SimMultiCorrData]{rcorrvar2}}).
    For an ordinal variable with r >= 2 categories, the variance is given by:
    \deqn{\sum_{j=1}^{r} {y_{j}}^{2}*p_{j} - {(\sum_{j=1}^{r} y_{j}*p_{j})}^{2}}.  Here, \eqn{y_{j}} is the j-th support
    value and \eqn{p_{j}} is \eqn{Pr(Y = y_{j})}.  This function would not ordinarily be called by the user.
}
\references{
Olsson U, Drasgow F, & Dorans NJ (1982). The Polyserial Correlation Coefficient. Psychometrika, 47(3): 337-47.
    \doi{10.1007/BF02294164}.
}
\seealso{
\code{\link[SimMultiCorrData]{ordnorm}}, \code{\link[SimMultiCorrData]{rcorrvar}},
    \code{\link[SimMultiCorrData]{rcorrvar2}}, \code{\link[SimMultiCorrData]{findintercorr_cont_cat}},
    \code{\link[SimMultiCorrData]{findintercorr_cont_pois2}}, \code{\link[SimMultiCorrData]{findintercorr_cont_nb2}}
}
\keyword{variance}
